core: New implementation for cgroup.rs
This commit replaces the legacy implementation of the cgroup module
found inside the core crate with a new implementation, that is more
centered around the tree approach found behind the cgroup architecture.
With this approach, it tries to simplifies things and remove
useless/obsolete functions found inside the legacy module.

Beside this, the new cgroup module also contains plenty of unit tests,
which however, cannot run in parallel, because they have a shared state
(the filesystem).