core: replace the cgroup Drop trait with rm()
This commit replaces the cgroup Drop trait with a separate method called
rm(). The change became necessary, because the initial approach makes it
almost impossible fetching the parents/children of a cgroup without
screwing everything up. Beside this, a Drop trait shall never fail,
which could have been possible in the previous implementation.