test: use sleep(1) instead of yes(1)
This commit changes the internal test function spawn_proc() inside the
cgroup.rs module of the core crate to call sleep(1) with a timeout
instead of yes(1). The reason for this is, that, in the case of a test
failure, the yes(1) process may become a zombie eating up system
resources. Using sleep(1) has the advantage that it will not be a zombie
for longer than timeout seconds.