core: remove CGroups by depth
This commit provides a partial fix for #17, by fixing the errors that
happened during the removal of a CGroup.

The fix works by obtaining the sub-directories of a CGroup in a sorted
manner, meaning that the outer-most directories are being removed before
the lower ones are. This step is necessary, because removing non-empty
directories is prohibited.

Unfortunately, the current approach requires the use of a capped busy
loop. Originally, it was planned to use a poller for that approach.
However, it quickly became clear that a poller would not provide the
correct solution for this problem, as pollers only indicate if data can
be read from or written to a file descriptor. In the case of our file
descriptor, a regular file on the filesystem, the read condition would
have been always true.

Instead, we thought about using the inotifiy kernel mechanism for this
problem, which issues a notification to the interfacing application,
whenever a certain file on the filesystem gets modified. However, this
kernel facility only works with normal files on normal filesystem,
thereby excluding virtual filesystems such as `/proc` and `/sys`.
Because of this, a busy loop is inavoidable.