core: introduce veth(4) pairs for network access
This commit introduces the net module inside the core crate, which
sondemly serves the purpose of creating veth(4) pairs for networking
access from the host system to the partition.

Currently, the creation of veth(4) interfaces is done through an
invocation of the ip(8) command, which is ugly. In future, things may
change, if a sane crate for high-level rtnetlink(7) operations appears.

Hypervisors can be configured to create veth(4) pairs, by specifying an
"interface" field inside the hypervisor config for a partition. Each
entry looks like this "p1:p2", whereas p1 will be created inside the
networking namespace of the interfacing application, and p2 will be
moved inside the namespace of the accompanying partition.