cgroup: use the events file for population check
This commit changes the logic of the CGroup::populated() function, in
order to use the cgroup.events file, rather than the cgorup.procs file
in order to check for the population. The reason for this is, that it is
faster reading a few bytes small file, rather than a file with an
theoretical unlimited size, which is the case for the cgroup.procs file,
due to the lack of a limit for processes within a cgroup.