core: revert the cgroup rewrite
This commit reverts the cgroup rewrite performed in #12, because it apparently
was not tested a lot. Most pull requests and bug reports in the past few months
were related to the new cgroup module. It even led to the lack of execution for
the periodic process.

While the legacy implementation is not perfect either, it is still better than
the current one, because it's author had a wider overview on what was going on.