Import process cgroup root inside partition
This imports the existing root cgroup of the partition processes instead of
recreating them. The cgroup root of a partition process is always created when
a partition is created. Attempting to create it again inside the partition is
therefore redundant and triggers a warning. As per the documentation, recreating
an existing cgroup root should always fail, because it indicates a logic error.
Therefore, this commit also changes `new_root` to treat recreating an existing
root as an error.