Prevent crash when logging many messages
The logger of a partition process uses Unix domain sockets to write messages to the hypervisor process.
This may fail temporarily, if the socket cannot receive any more messages.
The APEX does not provide a status code for signaling this condition to the partition process and instead the way to deal with any error when writing to the logger was to panic inside the partition process.

This commit prevents a crash when the socket can temporarily not be written to and instead ignores the error.
The tradeoff is that if the partition produces a great number of messages, some of them may not be logged.