move to workspace dependencies for a653rs and log
This change makes it easier to ensure that all members of the workspace
use the same version of a653rs and log, while still allowing exact
pinpointing of both. The pinpointing is important to avoid stale forks
to break when breaking changes in a653rs are released, which otherwise
easily slips through when using `version = "*"`.