remove device field from the partition config
This field was not used at all in the codebase, having it in the config
is misleading. Further on, it seems that there is no point in having it:
device files can be mounted through the mount option already.

Maybe we can eventually build something where for each partition
with at least one device mount we run `udevadm wait --settle <device>
[<device>]...` in a satellite thread, and we re-mount a device when it
pops back. That way we can re-attach a device without restarting the
hypervisor.

Further discussion in #80