fix(file_mounter): remove hidden conversion of absolute target path to relative path
If a FileMounter is initialized with an absolute target path, this path is converted to a relative path by stripping the first '/'. This is a hidden conversion not documented anywhere, which may lead to unexpected behaviour. The conversion is now replaced with an error stating that the target directory must be relative.